feat: drop ncurses, title features
