+++ configure.ac
@@ -1,5 +1,5 @@
 # Define the package name and version
-AC_INIT([koliseo], [0.4.5], [jgabaut@github.com])
+AC_INIT([koliseo], [0.4.6-dev], [jgabaut@github.com])
 
 # Verify automake version and enable foreign option
 AM_INIT_AUTOMAKE([foreign -Wall])
@@ -19,10 +19,6 @@ AC_ARG_ENABLE([debug],
   [AS_HELP_STRING([--enable-debug], [Enable debug build])],
   [enable_debug=$enableval],
   [enable_debug=no])
-AC_ARG_ENABLE([curses],
-  [AS_HELP_STRING([--enable-curses], [Enable curses functions])],
-  [enable_curses=$enableval],
-  [enable_curses=no])
 AC_ARG_ENABLE([gulp],
   [AS_HELP_STRING([--enable-gulp], [Enable gulp header])],
   [enable_gulp=$enableval],
@@ -41,7 +37,6 @@ AC_ARG_ENABLE([exper],
   [enable_exper=no])
 AM_CONDITIONAL([LOCATE_BUILD], [test "$enable_locate" = "yes"])
 AM_CONDITIONAL([DEBUG_BUILD], [test "$enable_debug" = "yes"])
-AM_CONDITIONAL([CURSES_BUILD], [test "$enable_curses" = "yes"])
 AM_CONDITIONAL([GULP_BUILD], [test "$enable_gulp" = "yes"])
 AM_CONDITIONAL([REGION_BUILD], [test "$enable_region" = "yes"])
 AM_CONDITIONAL([TITLE_BUILD], [test "$enable_title" = "yes"])
@@ -52,15 +47,9 @@ case "${host_os}" in
     mingw*)
         echo "Building for mingw32: [$host_cpu-$host_vendor-$host_os]"
         # mingw32 specific flags
-        if test "$enable_curses" = "yes"; then
-            echo "Building with curses header"
-            build_windows=yes
-            AC_SUBST([KOLISEO_CFLAGS], ["-I/usr/x86_64-w64-mingw32/include -static -fstack-protector -DNCURSES_STATIC"])
-            AC_SUBST([KOLISEO_LDFLAGS], ["-L/usr/x86_64-w64-mingw32/lib -lncursesw"])
-        else
-            AC_SUBST([KOLISEO_CFLAGS], ["-I/usr/x86_64-w64-mingw32/include -static -fstack-protector"])
-            AC_SUBST([KOLISEO_LDFLAGS], ["-L/usr/x86_64-w64-mingw32/lib"])
-        fi
+        build_windows=yes
+        AC_SUBST([KOLISEO_CFLAGS], ["-I/usr/x86_64-w64-mingw32/include -static -fstack-protector"])
+        AC_SUBST([KOLISEO_LDFLAGS], ["-L/usr/x86_64-w64-mingw32/lib"])
         AC_SUBST([CCOMP], ["/usr/bin/x86_64-w64-mingw32-gcc"])
         AC_SUBST([OS], ["w64-mingw32"])
         AC_SUBST([TARGET], ["demo.exe"])
@@ -70,14 +59,8 @@ case "${host_os}" in
         echo "Building for macos: [$host_cpu-$host_vendor-$host_os]"
         build_mac=yes
         # macOS specific flags
-        if test "$enable_curses" = "yes"; then
-            echo "Building with curses header"
-            AC_SUBST([KOLISEO_LDFLAGS], ["-L/opt/homebrew/opt/ncurses/lib -lncurses"])
-            AC_SUBST([KOLISEO_CFLAGS], ["-I/opt/homebrew/opt/ncurses/include"])
-        else
-            AC_SUBST([KOLISEO_LDFLAGS], [""])
-            AC_SUBST([KOLISEO_CFLAGS], [""])
-        fi
+        AC_SUBST([KOLISEO_LDFLAGS], [""])
+        AC_SUBST([KOLISEO_CFLAGS], [""])
         AC_SUBST([OS], ["darwin"])
         AC_SUBST([TARGET], ["demo"])
         AC_SUBST([SHARED_LIB], ["libkoliseo.so"])
@@ -86,12 +69,7 @@ case "${host_os}" in
         echo "Building for Linux: [$host_cpu-$host_vendor-$host_os]"
         build_linux=yes
         # Linux specific flags
-        if test "$enable_curses" = "yes"; then
-            echo "Building with curses header"
-            AC_SUBST([KOLISEO_LDFLAGS], ["-lncurses"])
-        else
-            AC_SUBST([KOLISEO_LDFLAGS], [""])
-        fi
+        AC_SUBST([KOLISEO_LDFLAGS], [""])
         AC_SUBST([KOLISEO_CFLAGS], [""])
         AC_SUBST([OS], ["Linux"])
         AC_SUBST([TARGET], ["demo"])
@@ -109,7 +87,7 @@ AM_CONDITIONAL([LINUX_BUILD], [test "$build_linux" = "yes"])
 # Set a default version number if not specified externally
 AC_ARG_VAR([VERSION], [Version number])
 if test -z "$VERSION"; then
-  VERSION="0.4.5"
+  VERSION="0.4.6-dev"
 fi
 
 # Output variables to the config.h header
