+++ docs/koliseo.doxyfile
@@ -48,7 +48,7 @@ PROJECT_NAME           = koliseo
 # could be handy for archiving the generated documentation or if some version
 # control system is used.
 
-PROJECT_NUMBER         = 0.4.5
+PROJECT_NUMBER         = 0.4.6-dev
 
 # Using the PROJECT_BRIEF tag one can provide an optional one line description
 # for a project that appears at the top of each page and should give viewer a
