+++ src/koliseo.c
@@ -69,74 +69,6 @@ const char* kls_reglist_backend_strings[KLS_REGLIST_TOTAL_BACKENDS] = {
 
 static bool kls_set_conf(Koliseo * kls, KLS_Conf conf);	//Declare function used internally by kls_new() and kls_new_conf()
 
-#ifdef KOLISEO_HAS_TITLE
-/**
- * Defines title banner.
- */
-char *kls_title[KLS_TITLEROWS + 1] = {
-    "                               .',,,.                                                               ",
-    "                        ...''',,;;:cl;.                                                             ",
-    "                  ..''''''....      .co,                                                            ",
-    "              .'',,,'..               'ol.                                                          ",
-    "          ..''''..                     .co.                                                         ",
-    "        ..'..                            cl.                                                        ",
-    "      ..'...               .          .  .:'...       .       .      ..      .....     ...          ",
-    "     ....      .      ..:c,'.   .oc .ll.  :dddxo.    ld.     'x:  .cxddx;   :kxodo,  .lddxx;        ",
-    "    .,..     .cl,..   .cxd:..   ;Ol;do.  :Ol. c0c   .kx.     lO;  ;0d..::  .xk'     .xx' 'kx.       ",
-    "    .,..     'dOl.     .''......lKOkc.  .kx.  ;0l   ;0l     .xk.  'xk;     'Ok;'.   l0;  .xk.       ",
-    "    ''.;,    .,;;............. .xKKO'   ;0c   c0:   lO;     'Od.   .lko.   :0koo:. .xk.  'Od.       ",
-    "   ',.'c;.  .......            'Oo:ko.  c0;  .xk.  .xx.     :0c  .   ;Od. .oO,     'Od.  c0:        ",
-    "   ,;  .......        .::.     :O; lO;  :0l..okc.  'Od...  .oO' .ld'.:Od. 'kk,...  .kk,.:kd.        ",
-    "   ;xc,...  .   'd:   .:d;     ;l. .l:  .:dool''c. 'ddolc. .cc.  'ldooc.  'dxoll:.  'odoo:.         ",
-    "  .lx;. .  ,d;  .ll.   .''.                     ;l:'                                                ",
-    "  :o.   ;,  cc    ..     .                       ,cc:::c,                                           ",
-    " .c;    ..  .'           ....',,;;;;;;;;;;,,,,,,,,,,;;;ox;........'cddoollcc:;,,'..                 ",
-    "  ,:         ...',:clodxkO0KKXXXXXXKKK000000KKKKXXXXKKKXXXKKKKKKKKKXNNNNNNNNNNNXXKOxoc;'.           ",
-    "  .:. ..';cldkOKXXXXK0Okxxdolc::;,''.','......';clc::cloONNNNNNNNNNNNNNNNNNNNNNNNNNNNNNK0xl;.       ",
-    "  :xxxOKKK0kxoddl;,';od;.   'cl,    .ckko.    ,d00Ol.   .xXNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNKkl;.   ",
-    ".lKX0xoc,';;. .ll.   ;xO;   .oX0,    ,ONXl    ,0NNNK;    .lKNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNX0o' ",
-    ".xk;.  .  .;.  .:,    .dc    'OK;    .dNNo.   ;0NNNXc      ;kXNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNk.",
-    " :c        ..   ..    .l,    .k0,    .xXXo    ;0XNNXc       .l0NNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNO'",
-    " :;                   .'.    .;:.    .,cl;....,loddo;..       'dKNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNN0,",
-    ".c,        ....',;;:cclllooddddddddddodxxxxxxxdddddddolllccccc:cxXNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNK;",
-    ".l:',:cloxxkkkOkkxxdollcc::;,,'''.............................',cONNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNXc",
-    ":00OOxdoc:;,'...                     ..        ..         .     .oNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNXl",
-    ":o,..',. ..   ';.   .co'     ;d,    'oOk;    .o00kd;    .oOkd,   :KNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNXl",
-    ":;   ,o'  ..  .l:    ;0o    .dXo.   'xNNk.   'ONNNNO'   :KNNNd.   cKNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNXo",
-    "c;    ..  ..   ';    'ko.   'ONo.   .kNNO'   ,0NNNNK;   :KNNXd.    :0NNNNNNNNNNNNNNNNNNNNNNNNNNNNNXo",
-    "l,             ..    'ko    'OXl    ,0NNO'   'ONNNNK;   ;KNNXl.     ;ONNNNNNNNNNNNNNNNNNNNNNNNNNNNXl",
-    "cc.............';'...;do'...:k0c....:0XX0:...'xXXXX0c...:0XK0l.......c0NNNNNNNNNNNNNNNNNNNNXXXXXXKO,",
-    ".;:;;;;;;:::::::::cc:::::::::c:;;;;,;ccc:;,,,,:cccc:;;;;;ccc:;,,,,,;;;clloooooooooooooooooollcc:;,. "
-};
-
-/**
- * Prints the title banner to the passed FILE pointer.
- * @see kls_title
- * @param fp The FILE to print to.
- */
-void kls_print_title_2file(FILE *fp)
-{
-    if (fp == NULL) {
-        fprintf(stderr,
-                "[KLS] kls_print_title_2file():  Passed file pointer was NULL.\n");
-        exit(EXIT_FAILURE);
-    }
-    for (int i = 0; i < KLS_TITLEROWS; i++) {
-        fprintf(fp, "%s\n", kls_title[i]);
-    }
-}
-
-/**
- * Prints the title banner to stdout.
- * @see kls_title
- */
-void kls_print_title(void)
-{
-    kls_print_title_2file(stdout);
-}
-
-#endif // KOLISEO_HAS_TITLE
-
 /**
  * Returns the constant string representing current version for Koliseo.
  * @return A constant string in MAJOR-MINOR-PATCH format for current Koliseo version.
@@ -210,11 +142,6 @@ void kls_dbg_features(void)
 #else
     bool kls_locate = false;
 #endif
-#ifdef KOLISEO_HAS_CURSES
-    bool kls_ncurses = true;
-#else
-    bool kls_ncurses = false;
-#endif
 #ifdef KOLISEO_HAS_GULP
     bool kls_gulp = true;
 #else
@@ -224,11 +151,6 @@ void kls_dbg_features(void)
     bool kls_region = true;
 #else
     bool kls_region = false;
-#endif
-#ifdef KOLISEO_HAS_TITLE
-    bool kls_title = true;
-#else
-    bool kls_title = false;
 #endif
 #ifdef KLS_DEBUG_CORE
     bool kls_debug = true;
@@ -240,17 +162,15 @@ void kls_dbg_features(void)
 #else
     bool kls_exper = false;
 #endif
-    bool features[7] = {
+    bool features[5] = {
         [0] = kls_debug,
         [1] = kls_locate,
-        [2] = kls_title,
-        [3] = kls_ncurses,
-        [4] = kls_region,
-        [5] = kls_gulp,
-        [6] = kls_exper,
+        [2] = kls_region,
+        [3] = kls_gulp,
+        [4] = kls_exper,
     };
     int total_enabled = 0;
-    for (int i=0; i<7; i++) {
+    for (int i=0; i<5; i++) {
         if (features[i]) {
             total_enabled += 1;
         }
@@ -266,14 +186,6 @@ void kls_dbg_features(void)
         }
         if (kls_locate) {
             fprintf(stderr, "locate%s", (total_enabled > 1 ? ", " : ""));
-            total_enabled -= 1;
-        }
-        if (kls_title) {
-            fprintf(stderr, "title%s", (total_enabled > 1 ? ", " : ""));
-            total_enabled -= 1;
-        }
-        if (kls_ncurses) {
-            fprintf(stderr, "ncurses%s", (total_enabled > 1 ? ", " : ""));
             total_enabled -= 1;
         }
         if (kls_region) {
@@ -1875,373 +1787,6 @@ void kls_formatSize(ptrdiff_t size, char *outputBuffer, size_t bufferSize)
 
     snprintf(outputBuffer, bufferSize, "%.2f %s", sizeValue, units[unitIndex]);
 }
-
-#ifdef KOLISEO_HAS_CURSES
-/**
- * Prints fields and eventually KLS_Region_List from the passed Koliseo pointer, to the passed WINDOW pointer.
- * @param kls The Koliseo at hand.
- * @param win The Window at hand.
- */
-void kls_show_toWin(Koliseo *kls, WINDOW *win)
-{
-    if (kls == NULL) {
-        fprintf(stderr, "kls_show_toWin(): passed Koliseo was null.");
-        exit(EXIT_FAILURE);
-    }
-    if (win == NULL) {
-#ifdef KLS_DEBUG_CORE
-        kls_log(kls, "ERROR", "kls_show_toWin():  passed WINDOW was null.");
-#else
-        fprintf(stderr, "kls_show_toWin(): passed WINDOW was null.");
-#endif
-        exit(EXIT_FAILURE);
-    }
-    wclear(win);
-    box(win, 0, 0);
-    wrefresh(win);
-    int y = 2;
-    int x = 2;
-    mvwprintw(win, y++, x, "Koliseo data:");
-    mvwprintw(win, y++, x, "API Level: { %i }", int_koliseo_version());
-#ifndef _WIN32
-    mvwprintw(win, y++, x, "Size: { %li }", kls->size);
-#else
-    mvwprintw(win, y++, x, "Size: { %lli }", kls->size);
-#endif
-    char h_size[200];
-    kls_formatSize(kls->size, h_size, sizeof(h_size));
-    mvwprintw(win, y++, x, "Human size: { %s }", h_size);
-    char curr_size[200];
-    kls_formatSize(kls->offset, curr_size, sizeof(curr_size));
-    mvwprintw(win, y++, x, "Used (Human): { %s }", curr_size);
-#ifndef _WIN32
-    mvwprintw(win, y++, x, "Offset: { %li }", kls->offset);
-#else
-    mvwprintw(win, y++, x, "Offset: { %lli }", kls->offset);
-#endif
-#ifndef _WIN32
-    mvwprintw(win, y++, x, "Prev_Offset: { %li }", kls->prev_offset);
-#else
-    mvwprintw(win, y++, x, "Prev_Offset: { %lli }", kls->prev_offset);
-#endif
-#ifdef KOLISEO_HAS_REGION
-    mvwprintw(win, y++, x, "KLS_Region_List len: { %i }",
-              kls_rl_length(kls->regs));
-#endif
-    mvwprintw(win, y++, x, "Current usage: { %.3f%% }",
-              (kls->offset * 100.0) / kls->size);
-    mvwprintw(win, y++, x, "%s", "");
-    mvwprintw(win, y++, x, "q or Enter to quit.");
-    /*
-       KLS_Region_List rl = kls_copy(kls->regs);
-       while (!kls_empty(rl)) {
-       mvwprintw(win, y, x, "Prev_Offset: [%i]",kls->prev_offset);
-       }
-     */
-    wrefresh(win);
-    int ch = '?';
-    int quit = -1;
-    do {
-        quit = 0;
-        ch = wgetch(win);
-        switch (ch) {
-        case 10:
-        case 'q': {
-            quit = 1;
-        }
-        break;
-        default: {
-            quit = 0;
-        }
-        break;
-        }
-    } while (!quit);
-}
-
-/**
- * Takes a Koliseo_Temp pointer and prints fields and eventually KLS_Region_List from the referred Koliseo pointer, to the passed WINDOW pointer.
- * @param t_kls The Koliseo_Temp at hand.
- * @param win The Window at hand.
- */
-void kls_temp_show_toWin(Koliseo_Temp *t_kls, WINDOW *win)
-{
-    if (t_kls == NULL) {
-        fprintf(stderr, "kls_temp_show_toWin(): passed Koliseo_Temp was null.");
-        exit(EXIT_FAILURE);
-    }
-    Koliseo *kls = t_kls->kls;
-    if (kls == NULL) {
-        fprintf(stderr, "kls_temp_show_toWin(): referred Koliseo was null.");
-        exit(EXIT_FAILURE);
-    }
-    if (win == NULL) {
-#ifdef KLS_DEBUG_CORE
-        kls_log(kls, "ERROR",
-                "kls_temp_show_toWin():  passed WINDOW was null.");
-#else
-        fprintf(stderr, "kls_temp_show_toWin(): passed WINDOW was null.");
-#endif
-        exit(EXIT_FAILURE);
-    }
-    wclear(win);
-    box(win, 0, 0);
-    wrefresh(win);
-    int y = 2;
-    int x = 2;
-    mvwprintw(win, y++, x, "Koliseo_Temp data:");
-    mvwprintw(win, y++, x, "API Level: { %i }", int_koliseo_version());
-#ifndef _WIN32
-    mvwprintw(win, y++, x, "Temp Size: { %li }", kls->size - t_kls->offset);
-    mvwprintw(win, y++, x, "Refer Size: { %li }", kls->size);
-#else
-    mvwprintw(win, y++, x, "Temp Size: { %lli }", kls->size - t_kls->offset);
-    mvwprintw(win, y++, x, "Refer Size: { %lli }", kls->size);
-#endif
-    char h_size[200];
-    char curr_size[200];
-    kls_formatSize(kls->size - t_kls->offset, h_size, sizeof(h_size));
-    mvwprintw(win, y++, x, "Temp Human size: { %s }", h_size);
-    kls_formatSize(kls->size, h_size, sizeof(h_size));
-    mvwprintw(win, y++, x, "Inner Human size: { %s }", h_size);
-    kls_formatSize(kls->offset, curr_size, sizeof(curr_size));
-    mvwprintw(win, y++, x, "Inner Used (Human): { %s }", curr_size);
-    kls_formatSize(t_kls->offset, curr_size, sizeof(curr_size));
-    mvwprintw(win, y++, x, "Temp Used (Human): { %s }", curr_size);
-#ifndef _WIN32
-    mvwprintw(win, y++, x, "Inner Offset: { %li }", kls->offset);
-    mvwprintw(win, y++, x, "Temp Offset: { %li }", t_kls->offset);
-#else
-    mvwprintw(win, y++, x, "Inner Offset: { %lli }", kls->offset);
-    mvwprintw(win, y++, x, "Temp Offset: { %lli }", t_kls->offset);
-#endif
-#ifndef _WIN32
-    mvwprintw(win, y++, x, "Inner Prev_Offset: { %li }", kls->prev_offset);
-    mvwprintw(win, y++, x, "Temp Prev_Offset: { %li }", t_kls->prev_offset);
-#else
-    mvwprintw(win, y++, x, "Inner Prev_Offset: { %lli }", kls->prev_offset);
-    mvwprintw(win, y++, x, "Temp Prev_Offset: { %lli }", t_kls->prev_offset);
-#endif
-#ifdef KOLISEO_HAS_REGION
-    mvwprintw(win, y++, x, "Refer KLS_Region_List len: { %i }",
-              kls_rl_length(kls->regs));
-    mvwprintw(win, y++, x, "Temp KLS_Region_List len: { %i }",
-              kls_rl_length(t_kls->t_regs));
-#endif
-    mvwprintw(win, y++, x, "Current inner usage: { %.3f%% }",
-              (kls->offset * 100.0) / kls->size);
-    mvwprintw(win, y++, x, "Current refer usage: { %.3f%% }",
-              (t_kls->offset * 100.0) / kls->size);
-    mvwprintw(win, y++, x, "%s", "");
-    mvwprintw(win, y++, x, "q or Enter to quit.");
-    /*
-       KLS_Region_List rl = kls_copy(kls->regs);
-       while (!kls_empty(rl)) {
-       mvwprintw(win, y, x, "Prev_Offset: [%i]",kls->prev_offset);
-       }
-     */
-    wrefresh(win);
-    int ch = '?';
-    int quit = -1;
-    do {
-        quit = 0;
-        ch = wgetch(win);
-        switch (ch) {
-        case 10:
-        case 'q': {
-            quit = 1;
-        }
-        break;
-        default: {
-            quit = 0;
-        }
-        break;
-        }
-    } while (!quit);
-}
-
-#ifdef KOLISEO_HAS_REGION
-/**
- * Displays a slideshow of KLS_Region_List from passed Koliseo, to the passed WINDOW pointer.
- * @param kls The Koliseo at hand.
- * @param win The Window at hand.
- */
-void kls_showList_toWin(Koliseo *kls, WINDOW *win)
-{
-    if (kls == NULL) {
-        fprintf(stderr, "kls_showList_toWin(): passed Koliseo was null.");
-        exit(EXIT_FAILURE);
-    }
-    if (win == NULL) {
-#ifdef KLS_DEBUG_CORE
-        kls_log(kls, "ERROR", "kls_showList_toWin():  passed WINDOW was null.");
-#else
-        fprintf(stderr, "kls_showList_toWin(): passed WINDOW was null.");
-#endif
-        exit(EXIT_FAILURE);
-    }
-    wclear(win);
-    box(win, 0, 0);
-    wrefresh(win);
-    int y = 2;
-    int x = 2;
-    int quit = 0;
-    mvwprintw(win, y++, x, "KLS_Region_List data:");
-    KLS_Region_List rl = kls->regs;
-    while (!quit && !kls_rl_empty(rl)) {
-        wclear(win);
-        y = 3;
-        KLS_list_element e = kls_rl_head(rl);
-        mvwprintw(win, y++, x, "Name: { %s }", e->name);
-        mvwprintw(win, y++, x, "Desc: { %s }", e->desc);
-#ifndef _WIN32
-        mvwprintw(win, y++, x, "Offsets: { %li } -> { %li }", e->begin_offset,
-                  e->end_offset);
-        mvwprintw(win, y++, x, "Size: { %li }", e->size);
-        mvwprintw(win, y++, x, "Padding: { %li }", e->padding);
-#else
-        mvwprintw(win, y++, x, "Offsets: { %lli } -> { %lli }", e->begin_offset,
-                  e->end_offset);
-        mvwprintw(win, y++, x, "Size: { %lli }", e->size);
-        mvwprintw(win, y++, x, "Padding: { %lli }", e->padding);
-#endif
-        mvwprintw(win, y++, x, "KLS_Region_List len: { %i }",
-                  kls_rl_length(kls->regs));
-        mvwprintw(win, y++, x, "Current usage: { %.3f%% }",
-                  kls_usageShare(e, kls));
-        char h_size[200];
-        ptrdiff_t reg_size = e->end_offset - e->begin_offset;
-        kls_formatSize(reg_size, h_size, sizeof(h_size));
-        mvwprintw(win, y++, x, "Human size: { %s }", h_size);
-        mvwprintw(win, y++, x, "%s", "");
-        mvwprintw(win, y++, x, "q to quit, Right arrow to go forward.");
-        /*
-           KLS_Region_List rl = kls_copy(kls->regs);
-           while (!kls_empty(rl)) {
-           mvwprintw(win, y, x, "Prev_Offset: [%i]",kls->prev_offset);
-           }
-         */
-        box(win, 0, 0);
-        wrefresh(win);
-        int ch = '?';
-        int picked = -1;
-        do {
-            picked = 0;
-            ch = wgetch(win);
-            switch (ch) {
-            case KEY_RIGHT: {
-                rl = kls_rl_tail(rl);
-                picked = 1;
-            }
-            break;
-            case 'q': {
-                quit = 1;
-                picked = 1;
-            }
-            break;
-            default: {
-                picked = 0;
-            }
-            break;
-            }
-        } while (!quit && !picked);
-    }
-}
-
-/**
- * Displays a slideshow of KLS_Region_List from passed Koliseo_Temp, to the passed WINDOW pointer.
- * @param t_kls The Koliseo_Temp at hand.
- * @param win The Window at hand.
- */
-void kls_temp_showList_toWin(Koliseo_Temp *t_kls, WINDOW *win)
-{
-    if (t_kls == NULL) {
-        fprintf(stderr,
-                "kls_temp_showList_toWin(): passed Koliseo_Temp was null.\n");
-        exit(EXIT_FAILURE);
-    }
-    Koliseo *kls_ref = t_kls->kls;
-    if (kls_ref == NULL) {
-        fprintf(stderr,
-                "kls_temp_showList_toWin(): referred Koliseo was null.\n");
-        exit(EXIT_FAILURE);
-    }
-    if (win == NULL) {
-#ifdef KLS_DEBUG_CORE
-        kls_log(kls_ref, "ERROR",
-                "kls_temp_showList_toWin():  passed WINDOW was null.");
-#else
-        fprintf(stderr, "kls_temp_showList_toWin(): passed WINDOW was null.\n");
-#endif
-        exit(EXIT_FAILURE);
-    }
-    wclear(win);
-    box(win, 0, 0);
-    wrefresh(win);
-    int y = 2;
-    int x = 2;
-    int quit = 0;
-    mvwprintw(win, y++, x, "KLS_Region_List data:");
-    KLS_Region_List rl = t_kls->t_regs;
-    while (!quit && !kls_rl_empty(rl)) {
-        wclear(win);
-        y = 3;
-        KLS_list_element e = kls_rl_head(rl);
-        mvwprintw(win, y++, x, "Name: { %s }", e->name);
-        mvwprintw(win, y++, x, "Desc: { %s }", e->desc);
-#ifndef _WIN32
-        mvwprintw(win, y++, x, "Offsets: { %li } -> { %li }", e->begin_offset,
-                  e->end_offset);
-        mvwprintw(win, y++, x, "Size: { %li }", e->size);
-        mvwprintw(win, y++, x, "Padding: { %li }", e->padding);
-#else
-        mvwprintw(win, y++, x, "Offsets: { %lli } -> { %lli }", e->begin_offset,
-                  e->end_offset);
-        mvwprintw(win, y++, x, "Size: { %lli }", e->size);
-        mvwprintw(win, y++, x, "Padding: { %lli }", e->padding);
-#endif
-        mvwprintw(win, y++, x, "KLS_Region_List len: { %i }",
-                  kls_rl_length(t_kls->t_regs));
-        //mvwprintw(win, y++, x, "Current usage: { %.3f%% }", kls_usageShare(e,kls));
-        char h_size[200];
-        ptrdiff_t reg_size = e->end_offset - e->begin_offset;
-        kls_formatSize(reg_size, h_size, sizeof(h_size));
-        mvwprintw(win, y++, x, "Human size: { %s }", h_size);
-        mvwprintw(win, y++, x, "%s", "");
-        mvwprintw(win, y++, x, "q to quit, Right arrow to go forward.");
-        /*
-           KLS_Region_List rl = kls_copy(kls->regs);
-           while (!kls_empty(rl)) {
-           mvwprintw(win, y, x, "Prev_Offset: [%i]",kls->prev_offset);
-           }
-         */
-        box(win, 0, 0);
-        wrefresh(win);
-        int ch = '?';
-        int picked = -1;
-        do {
-            picked = 0;
-            ch = wgetch(win);
-            switch (ch) {
-            case KEY_RIGHT: {
-                rl = kls_rl_tail(rl);
-                picked = 1;
-            }
-            break;
-            case 'q': {
-                quit = 1;
-                picked = 1;
-            }
-            break;
-            default: {
-                picked = 0;
-            }
-            break;
-            }
-        } while (!quit && !picked);
-    }
-}
-#endif // KOLISEO_HAS_REGION
-#endif //KOLISEO_HAS_CURSES
 
 /**
  * Resets the offset field for the passed Koliseo pointer.
